==> colon_RNAseq/scripts/13_module_member_heatmaps.R <==
#This script ranks genes by their meta-module membership to their respective module
#Then, two plots are created:
    #1: a barchart showing the meta-module membership within only our dataset
    #2: a large heatmap showing the regional/dataset-specific module membership (kME) of those genes
#This plot was MAINLY generated to provide the reader with lists of interesting/relevant genes, but also may corroborate the module preservation analysis
